# tools/audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraderCopilot Backend",
    "version": "2.0.0"
  },
  "paths": {
...
    "/advisor/chat": {
      "post": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Advisor Chat",
        "description": "Endpoint for interactive chat with the Advisor.\nEnforces Entitlements:\n1. Plan Access (can_use_advisor)\n2. Daily Quota (advisor_chat)\n3. Token Access (if context token provided)",
        "operationId": "advisor_chat_advisor_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/advisor/": {
      "post": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Analyze Advisor",
        "description": "Analiza una posici\u00f3n abierta y sugiere alternativas.\nVersi\u00f3n local determinista (sin LLM).\nEnforces Entitlements.",
        "operationId": "analyze_advisor_advisor__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisorReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login For Access Token",
        "description": "Endpoint estandar OAuth2 para login (email/password).",
        "operationId": "login_for_access_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register User",
        "description": "(Dev Helper) Registra un usuario inicial.",
        "operationId": "register_user_auth_register_post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Trader",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/entitlements": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Read My Entitlements",
        "description": "Diagn\u00f3stico de l\u00edmites y cuotas.",
        "operationId": "read_my_entitlements_auth_me_entitlements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/analyze/advisor/chat": {
      "post": {
        "summary": "Analyze Advisor Chat",
        "description": "Endpoint para chat interactivo con el Advisor.\nEnforced Entitlements:\n- Check Feature Access (can_use_advisor).\n- Check Token Access (if context implies a specific token).\n- Check & Increment 'advisor_chat' Quota (if we wanted to limit messages, though plan said Unlimited for Pro).\n  But we should still check if they HAVE access.",
        "operationId": "analyze_advisor_chat_analyze_advisor_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisorChatReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvisorChatReq": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "History"
          },
          "context": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "history",
          "context"
        ],
        "title": "AdvisorChatReq"
      },
      "AdvisorReq": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Token base: eth|btc|sol|xau"
          },
          "direction": {
            "type": "string",
            "enum": [
              "long",
              "short"
            ],
            "title": "Direction",
            "description": "Direcci\u00f3n de la posici\u00f3n"
          },
          "entry": {
            "type": "number",
            "title": "Entry",
            "description": "Precio de entrada de la posici\u00f3n"
          },
          "tp": {
            "type": "number",
            "title": "Tp",
            "description": "Take profit previsto"
          },
          "sl": {
            "type": "number",
            "title": "Sl",
            "description": "Stop loss previsto"
          },
          "size_quote": {
            "type": "number",
            "title": "Size Quote",
            "description": "Tama\u00f1o de la posici\u00f3n en moneda de cotizaci\u00f3n (ej. USDT)"
          }
        },
        "type": "object",
        "required": [
          "token",
          "direction",
          "entry",
          "tp",
          "sl",
          "size_quote"
        ],
        "title": "AdvisorReq"
      },
...
      "Body_login_for_access_token_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_for_access_token_auth_token_post"
      },
      "ChatContext": {
        "properties": {
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframe"
          },
          "signal_data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Data"
          }
        },
        "type": "object",
        "title": "ChatContext"
      },
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}